add e2e tests for gitlab.com as slow tests which do not run by default
