+++ tests/gitlab_test_utils.py
@@ -1,4 +1,3 @@
-
 import pytest
 import json
 from unittest import mock
