+++ tests/io_test_util.py
@@ -1,3 +1,6 @@
+import os
+import sys
+import subprocess
 import sys
 from contextlib import contextmanager
 from io import StringIO
@@ -12,3 +15,11 @@ def captured_output():
         yield sys.stdout, sys.stderr
     finally:
         sys.stdout, sys.stderr = old_out, old_err
+
+def execute(args, timeout=3):
+    cmd = [sys.executable, '-m', 'gitlabber']
+    cmd.extend(args)
+    with subprocess.Popen(cmd, stdout=subprocess.PIPE, env=os.environ.copy()) as process:
+        outs, err = process.communicate(timeout=timeout)    
+        process.wait()
+        return outs.decode('utf-8')
